requirements.txt
Django==4.2
djangorestframework==3.14.0
celery==5.3.4
gunicorn==21.2.0
redis==5.0.1
django-redis==5.4.0
python-decouple==3.8.0
psycopg2-binary==2.9.9
mssql-django==1.4
django-webpack-loader
python-dotenv==1.0.0
django-bootstrap4==23.2
python-ldap==3.4.4
django-auth-ldap==4.6.0
pandas==2.1.4
openpyxl==3.1.5
xlrd==2.0.1
requests==2.32.3
weasyprint==62.3
